pyproject.toml
[tool.poetry]
name = "zeek"
version = "1.0.0"
description = "Discord server metric collection"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
loguru = "^0.5.3"
jishaku = "^1.20.0"
asyncpg = "^0.22.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}

[tool.poetry.dev-dependencies]
flake8 = "^3.8"
flake8-annotations = "^2.3"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.5"
flake8-import-order = "^0.18"
flake8-string-format = "^0.3"
flake8-tidy-imports = "^4.1"
flake8-todo = "^0.7"
pep8-naming = "^0.11"
pre-commit = "^2.1"
taskipy = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
